@msdev
https://digitalcommons.csumb.edu/cgi/viewcontent.cgi?article=1108&context=hornbeck_usa_3_d
https://archives.cdn.sos.ca.gov/collections/1849/images/1849Constitucion2.pdf
[Constitutions](https://www.sos.ca.gov/archives/collections/constitutions/1849) @stateofcalifornia @cityoflosangeles @nasa @nasa-jpl @whitehouse * CALIFORNIA CONSTITUTION - CONS

ARTICLE I DECLARATION OF RIGHTS [SECTION 1 - SEC. 32]  ( Article 1 adopted 1879. )
  
SEC. 2.  
(a) Every person may freely speak, write and publish his or her sentiments on all subjects, being responsible for the abuse of this right. A law may not restrain or abridge liberty of speech or press.

(b) A publisher, editor, reporter, or other person connected with or employed upon a newspaper, magazine, or other periodical publication, or by a press association or wire service, or any person who has been so connected or employed, shall not be adjudged in contempt by a judicial, legislative, or administrative body, or any other body having the power to issue subpoenas, for refusing to disclose the source of any information procured while so connected or employed for publication in a newspaper, magazine or other periodical publication, or for refusing to disclose any unpublished information obtained or prepared in gathering, receiving or processing of information for communication to the public.

Nor shall a radio or television news reporter or other person connected with or employed by a radio or television station, or any person who has been so connected or employed, be so adjudged in contempt for refusing to disclose the source of any information procured while so connected or employed for news or news commentary purposes on radio or television, or for refusing to disclose any unpublished information obtained or prepared in gathering, receiving or processing of information for communication to the public.

As used in this subdivision, “unpublished information” includes information not disseminated to the public by the person from whom disclosure is sought, whether or not related information has been disseminated and includes, but is not limited to, all notes, outtakes, photographs, tapes or other data of whatever sort not itself disseminated to the public through a medium of communication, whether or not published information based upon or related to such material has been disseminated.

[(Sec. 2 amended June 3, 1980, by Prop. 5. Res.Ch. 77, 1978.)  ](https://leginfo.legislature.ca.gov/faces/codes_displaySection.xhtml?lawCode=CONS&sectionNum=SEC.%202.&article=I)